refactor: move ui folder to  root of components folder
