+++ components/post-ad/PostAdForm.tsx
@@ -4,8 +4,8 @@ import React, { useState } from "react";
 import { SubmitErrorHandler, SubmitHandler, useForm } from "react-hook-form";
 import { z } from "zod";
 import InputField from "./InputField";
-import { Button } from "../auth/ui/button";
-import { Form } from "../auth/ui/form";
+import { Button } from "../ui/button";
+import { Form } from "../ui/form";
 import Selector from "./Selector";
 import AddPhoto from "./AddPhoto";
 import { Separator } from "@radix-ui/react-select";
